# gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  genprotos.AuthMessage:
    properties:
      message:
        type: string
    type: object
...
  genprotos.EditProfileRequest:
    properties:
      bio:
        type: string
      full_name:
        type: string
      id:
        type: string
    type: object
  genprotos.EditProfileResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      password:
        type: string
      user_type:
        type: string
      username:
        type: string
    type: object
  genprotos.EditUserTypeRequest:
    properties:
      id:
        type: string
      user_type:
        type: string
    type: object
  genprotos.EditUserTypeResponse:
    properties:
      id:
        type: string
      updated_at:
        type: string
      user_type:
        type: string
      username:
        type: string
    type: object
...
  genprotos.GetAllUsersResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/genprotos.User'
        type: array
    type: object
...
  genprotos.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  genprotos.LoginResponse:
    properties:
      "true":
        type: boolean
    type: object
  genprotos.Message:
    properties:
      message:
        type: string
    type: object
...
  genprotos.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      user_type:
        type: string
      username:
        type: string
    type: object
  genprotos.RegisterResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      password:
        type: string
      user_type:
        type: string
      username:
        type: string
    type: object
  genprotos.ResetPasswordRequest:
    properties:
      email:
        type: string
    type: object
...
  genprotos.ShowProfileResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      password:
        type: string
      updated_at:
        type: string
      user_type:
        type: string
      username:
        type: string
    type: object
...
  genprotos.User:
    properties:
      full_name:
        type: string
      id:
        type: string
      user_type:
        type: string
      username:
        type: string
    type: object
host: localhost:9090
info:
  contact: {}
  description: TEST
  title: API
  version: "1.0"
paths:
...
  /auth/delete/{id}:
    delete:
      consumes:
      - application/json
      description: This endpoint for deleting user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.AuthMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/genprotos.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/genprotos.Message'
      summary: Delete user
  /auth/login:
    post:
      consumes:
      - application/json
      description: This endpoint for logging in user.
      parameters:
      - description: User login details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genprotos.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/genprotos.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/genprotos.Message'
      summary: Login user
  /auth/profile/{id}:
    get:
      consumes:
      - application/json
      description: This endpoint for showing user profile.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.ShowProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/genprotos.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/genprotos.Message'
      summary: Show user profile
  /auth/profile/edit:
    put:
      consumes:
      - application/json
      description: This endpoint for editing user profile.
      parameters:
      - description: User profile details to edit
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genprotos.EditProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.EditProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/genprotos.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/genprotos.Message'
      summary: Edit user profile
  /auth/register:
    post:
      consumes:
      - application/json
      description: This endpoint for registering user.
      parameters:
      - description: User details to register
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genprotos.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/genprotos.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/genprotos.Message'
      summary: Register user
  /auth/reset:
    post:
      consumes:
      - application/json
      description: This endpoint for resetting user password.
      parameters:
      - description: User email to reset password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genprotos.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.AuthMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/genprotos.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/genprotos.Message'
      summary: Reset password
  /auth/users:
    get:
      consumes:
      - application/json
      description: This endpoint for getting all users with pagination.
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Number of users per page
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.GetAllUsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/genprotos.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/genprotos.Message'
      summary: Get all users
  /auth/usertype/edit:
    put:
      consumes:
      - application/json
      description: This endpoint for editing user type.
      parameters:
      - description: User type details to edit
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genprotos.EditUserTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.EditUserTypeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/genprotos.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/genprotos.Message'
      summary: Edit user type
...
swagger: "2.0"